requirements.txt
cycler==0.11.0
fonttools==4.34.2
future==0.18.2
iso8601==1.0.2
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.0
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pyserial==3.5
python-dateutil==2.8.2
PyWavelets==1.3.0
PyYAML==6.0
scipy==1.8.1
six==1.16.0
